Rename directories and endpoints
